requirements.txt
# Core dependencies
Flask==2.3.3
python-dotenv==1.0.0

# LINE Bot SDK
line-bot-sdk==3.5.0

# Microsoft Graph API
requests==2.31.0
msal==1.24.1

# OpenAI API
openai==1.3.5

# Database options
# Firebase
firebase-admin==6.2.0

# MySQL (uncomment if using MySQL)
# PyMySQL==1.1.0
# SQLAlchemy==2.0.23

# Utilities
Pillow==10.1.0
python-dateutil==2.8.2

# Development dependencies
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# Production server
gunicorn==21.2.0